Add code split in `ActionCollectionDTO` (#28557)
